+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2019-11-25
+
 ### Fixed
 - Return the tx-id instead of the tx-hash for Bitcoin because it is the canonical way of identifying a transaction.
 
@@ -63,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Common code that can be used to build applications on top of COMIT.
 
-[Unreleased]: https://github.com/comit-network/comit-js-sdk/compare/0.5.6...HEAD
+[Unreleased]: https://github.com/comit-network/comit-js-sdk/compare/0.6.0...HEAD
+[0.6.0]: https://github.com/comit-network/comit-js-sdk/compare/0.5.6...0.6.0
 [0.5.6]: https://github.com/comit-network/comit-js-sdk/compare/0.4.1...0.5.6
 [0.4.1]: https://github.com/comit-network/comit-js-sdk/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/comit-network/comit-js-sdk/compare/0.3.2...0.4.0